+++ src/alert/AlertList.vue
@@ -2,9 +2,9 @@
 import { watch } from "vue";
 import { useRoute } from "vue-router";
 import { PhX } from "@phosphor-icons/vue";
-import useMessenger from "@/message/messenger.composable";
+import useAlert from "@/alert/alert.composable";
 
-const { alerts, dismiss, clear } = useMessenger();
+const { alerts, dismiss, clear } = useAlert();
 const route = useRoute();
 
 // Remove messages when navigating to a new page
@@ -20,10 +20,9 @@ watch(
     class="container max-w-3xl my-4 sticky top-2 z-50"
   >
     <div
-      v-for="{ key, message, level } in alerts"
+      v-for="{ key, message } in alerts"
       :key="message"
-      class="opacity-80 rounded-lg my-2 text-white p-2 shadow-lg flex items-start"
-      :class="level == 'error' ? ['bg-red-500'] : ['bg-gray-400']"
+      class="opacity-80 rounded-lg my-2 bg-red-500 text-white p-2 shadow-lg flex items-start"
     >
       <div class="flex-1 px-2">
         {{ message }}