+++ src/exports/ExportViewTemplate.vue
@@ -6,7 +6,7 @@ import TextFileBox from "@/components/TextFileBox.vue";
 import LayoutSection from "@/components/LayoutSection.vue";
 import PendingContent from "@/spin/PendingContent.vue";
 import useLocale from "@/i18n/locale.composable";
-import useMessenger from "@/message/messenger.composable";
+import useAlert from "@/alert/alert.composable";
 import useExports from "@/exports/exports.composable";
 import type { ResourceType } from "@/api/api.types";
 import { useExportStore } from "@/store/export.store";
@@ -20,7 +20,7 @@ const props = defineProps<{
 const { loadExports } = useExportStore();
 const { loadResultFile } = useExports(props.type, props.id);
 const { filesize, formatDate } = useLocale();
-const { alert, alertError } = useMessenger();
+const { showAlert } = useAlert();
 const { t } = useI18n();
 
 const exports = computedAsync(() => loadExports(props.type, props.id));
@@ -33,12 +33,11 @@ const isXml = computed(() => /\/xml$/.test(metadata.value?.type || ""));
 
 // Show error if given filename is not found
 watchImmediate([exports, metadata], () => {
-  if (exports.value?.length && !metadata.value)
-    alert(t("source.notfound"), "error");
+  if (exports.value?.length && !metadata.value) showAlert(t("source.notfound"));
 });
 
 async function loadFile() {
-  return metadata.value && (await loadResultFile(path.value).catch(alertError));
+  return metadata.value && (await loadResultFile(path.value).catch(showAlert));
 }
 </script>